Moved web3 modules into web3 folder
